+++ Source/AnimatedImage.swift
@@ -3,73 +3,138 @@
 // Copyright (c) 2016 Alexander Grebenyuk (github.com/kean).
 
 import UIKit
+import ImageIO
 import FLAnimatedImage
 import Nuke
-import ImageIO
 
-/// Represents animated image.
+/// Represents animated image data alongside a poster image (first image frame).
 public class AnimatedImage: UIImage {
-    /// Image data that the receiver was initialized with.
-    public let data: NSData! // it's nonnull
+    public let data: Data
     
-    /// Initializes the receiver with a given data and poster image.
-    public init(data: NSData, poster: CGImageRef) {
+    public init(data: Data, poster: CGImage) {
         self.data = data
-        super.init(CGImage: poster, scale: 1, orientation: .Up)
+        super.init(cgImage: poster, scale: 1, orientation: .up)
     }
     
-    /// Not implemented.
-    public required init?(coder decoder: NSCoder) {
-        self.data = nil // makes me sad
-        super.init(coder: decoder)
+    public required init?(coder aDecoder: NSCoder) {
+        guard let data = aDecoder.decodeObject(forKey: "_nk_data") as? Data else {
+            return nil
+        }
+        self.data = data
+        super.init(coder: aDecoder)
     }
     
-    /// Not implemented.
-    public required convenience init(imageLiteral name: String) {
+    public override func encode(with aCoder: NSCoder) {
+        super.encode(with: aCoder)
+        aCoder.encode(data, forKey: "_nk_data")
+    }
+    
+    public required convenience init(imageLiteralResourceName name: String) {
         fatalError("init(imageLiteral:) has not been implemented")
     }
 }
 
-/// Creates instances of `AnimatedImage` class from the given data. Checks if the image data is in a GIF image format, otherwise returns nil.
-public class AnimatedImageDecoder: ImageDecoding {
-    /// Initializer the receiver.
+public extension AnimatedImage {
+    /// Default `Nuke.Manager` with animated GIF support.
+    public static let manager: Nuke.Manager = {
+        // Make a decoder which supports animated GIFs.
+        let decoder = Nuke.DataDecoderComposition(decoders: [AnimatedImageDecoder(), Nuke.DataDecoder()])
+        
+        // Updates `Cache` cost calculation block.
+        let cache = Nuke.Cache().preparedForAnimatedImages()
+        
+        let loader = Nuke.Loader(loader: Nuke.DataLoader(), decoder: decoder, cache: cache)
+        
+        // Disable processing of animated images.
+        loader.makeProcessor = { image, request in
+            return image is AnimatedImage ? nil : request.processor
+        }
+        return Manager(loader: loader, cache: cache)
+    }()
+}
+
+/// Creates instances of `AnimatedImage` class from the given data.
+/// Returns `nil` is data doesn't contain an animated GIF.
+public class AnimatedImageDecoder: Nuke.DataDecoding {
     public init() {}
     
-    /// Decodes image data and returns an instance of AnimatedImage class.
-    public func decode(data: NSData, response: NSURLResponse?) -> Image? {
+    public func decode(data: Data, response: URLResponse) -> Nuke.Image? {
         guard self.isAnimatedGIFData(data) else {
             return nil
         }
-        guard let poster = self.posterImageFor(data) else {
+        guard let poster = self.posterImage(for: data) else {
             return nil
         }
         return AnimatedImage(data: data, poster: poster)
     }
     
-    /// Return true if the image data represents animated image in GIF format.
-    public func isAnimatedGIFData(data: NSData) -> Bool {
+    public func isAnimatedGIFData(_ data: Data) -> Bool {
         let sigLength = 3
-        if data.length < sigLength {
+        if data.count < sigLength {
             return false
         }
-        var sig = [UInt8](count: sigLength, repeatedValue: 0)
-        data.getBytes(&sig, length:sigLength)
+        var sig = [UInt8](repeating: 0, count: sigLength)
+        (data as NSData).getBytes(&sig, length:sigLength)
         return sig[0] == 0x47 && sig[1] == 0x49 && sig[2] == 0x46
     }
     
-    private func posterImageFor(data: NSData) -> CGImageRef? {
-        if let source = CGImageSourceCreateWithData(data, nil) {
+    private func posterImage(for data: Data) -> CGImage? {
+        if let source = CGImageSourceCreateWithData(data as CFData, nil) {
             return CGImageSourceCreateImageAtIndex(source, 0, nil)
         }
         return nil
     }
-    
 }
 
-/// Extension that adds image loading capabilities to the FLAnimatedImageView.
+/// Simple `FLAnimatedImage` wrapper that implement `Nuke.Target` protocol.
+///
+/// The reason why this is a standalone class and not a simple overridden method
+/// from `UIImageView` extension is because declarations from extensions
+/// cannot be overridden in Swift (yet).
+public class AnimatedImageView: UIView, Nuke.Target {
+    public let imageView: FLAnimatedImageView
+    
+    public init(imageView: FLAnimatedImageView = FLAnimatedImageView()) {
+        self.imageView = imageView
+        
+        super.init(frame: CGRect.zero)
+        
+        addSubview(imageView)
+        
+        layoutMargins = UIEdgeInsets.zero
+        
+        imageView.translatesAutoresizingMaskIntoConstraints = false
+        for attr in [.top, .leading, .bottom, .trailing] as [NSLayoutAttribute] {
+            addConstraint(NSLayoutConstraint(item: imageView, attribute: attr, relatedBy: .equal, toItem: self, attribute: attr, multiplier: 1, constant: 0))
+        }
+    }
+    
+    required public init?(coder aDecoder: NSCoder) {
+        fatalError("AnimatedImageView doesn't support NSCoding yet")
+    }
+    
+    /// Displays an image on success. Runs `opacity` transition if
+    /// the response was not from the memory cache.
+    public func handle(response: Response, isFromMemoryCache: Bool) {
+        switch response {
+        case let .fulfilled(image):
+            imageView.nk_display(image)
+            if !isFromMemoryCache {
+                let animation = CABasicAnimation(keyPath: "opacity")
+                animation.duration = 0.25
+                animation.fromValue = 0
+                animation.toValue = 1
+                let layer: CALayer? = imageView.layer // Make compiler happy on OSX
+                layer?.add(animation, forKey: "imageTransition")
+            }
+        case .rejected(_): return
+        }
+    }
+}
+
 public extension FLAnimatedImageView {
     /// Displays a given image. Starts animation if image is an instance of AnimatedImage.
-    public override func nk_displayImage(image: Image?) {
+    public func nk_display(_ image: Image?) {
         guard image != nil else {
             self.animatedImage = nil
             self.image = nil
@@ -78,12 +143,14 @@ public extension FLAnimatedImageView {
         if let image = image as? AnimatedImage {
             // Display poster image immediately
             self.image = image
-
-            // Start playback after we prefare FLAnimatedImage for rendering
-            dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_DEFAULT, 0)) {
+            
+            // Prepare FLAnimatedImage object asynchronously (it takes a
+            // noticeable amount of time), and start playback.
+            DispatchQueue.global().async {
                 let animatedImage = FLAnimatedImage(animatedGIFData: image.data)
-                dispatch_async(dispatch_get_main_queue()) {
-                    if self.image === image { // Still displaying the same poster image
+                DispatchQueue.main.async {
+                    // If view is still displaying the same image
+                    if self.image === image {
                         self.animatedImage = animatedImage
                     }
                 }
@@ -94,33 +161,17 @@ public extension FLAnimatedImageView {
     }
 }
 
-/// Prevents `ImageLoader` from processing animated images.
-public class AnimatedImageLoaderDelegate: ImageLoaderDefaultDelegate {
-    /// Disabled processing of animated images.
-    public override func loader(loader: ImageLoader, processorFor request: ImageRequest, image: Image) -> ImageProcessing? {
-        return image is AnimatedImage ? nil : super.loader(loader, processorFor: request, image: image)
-    }
-}
-
-/// Memory cache that is aware of animated images. Can be used for both single-frame and animated images.
-public class AnimatedImageMemoryCache: ImageMemoryCache {
-
-    /// Can be used to disable storing animated images. Default value is true (storage is allowed).
-    public var allowsAnimatedImagesStorage = true
-
-    /// Stores response unless the image is an instance of AnimatedImage class and animated image storage is disabled.
-    public override func setResponse(response: ImageCachedResponse, forKey key: ImageRequestKey) {
-        if !self.allowsAnimatedImagesStorage && response.image is AnimatedImage {
-            return
+public extension Nuke.Cache {
+    /// Updates `Cache` cost block by adding special handling of `AnimatedImage`.
+    public func preparedForAnimatedImages() -> Self {
+        let cost = self.cost
+        self.cost = {
+            var val = cost($0)
+            if let animatedImage = $0 as? AnimatedImage {
+                val += animatedImage.data.count
+            }
+            return val
         }
-        super.setResponse(response, forKey: key)
-    }
-
-    /// Returns cost for a given image.
-    public override func costFor(image: Image) -> Int {
-        if let animatedImage = image as? AnimatedImage {
-            return animatedImage.data.length + super.costFor(image)
-        }
-        return super.costFor(image)
+        return self
     }
 }